# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SceneDream",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "description": "Return global defaults and active art-style catalog entries.",
        "operationId": "settings-get_settings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsBundleResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Update Settings",
        "description": "Update global defaults for scenes-per-run and prompt art style.",
        "operationId": "settings-update_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsBundleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/art-style-lists": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Art Style Lists",
        "description": "Return recommended/other style pools as line-oriented settings arrays.",
        "operationId": "settings-get_art_style_lists",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtStyleListsRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Art Style Lists",
        "description": "Replace active recommended/other style pools from full list payloads.",
        "operationId": "settings-update_art_style_lists",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtStyleListsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtStyleListsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppSettingsBundleResponse": {
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/AppSettingsRead"
          },
          "art_styles": {
            "items": {
              "$ref": "#/components/schemas/ArtStyleRead"
            },
            "type": "array",
            "title": "Art Styles"
          }
        },
        "type": "object",
        "required": [
          "settings",
          "art_styles"
        ],
        "title": "AppSettingsBundleResponse",
        "description": "Settings payload bundled with active art styles."
      },
      "AppSettingsRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "default_scenes_per_run": {
            "type": "integer",
            "title": "Default Scenes Per Run"
          },
          "default_prompt_art_style_mode": {
            "type": "string",
            "enum": [
              "random_mix",
              "single_style"
            ],
            "title": "Default Prompt Art Style Mode"
          },
          "default_prompt_art_style_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Prompt Art Style Text"
          },
          "social_posting_enabled": {
            "type": "boolean",
            "title": "Social Posting Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "default_scenes_per_run",
          "default_prompt_art_style_mode",
          "default_prompt_art_style_text",
          "social_posting_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "AppSettingsRead",
        "description": "Detailed representation of current global settings."
      },
      "AppSettingsUpdateRequest": {
        "properties": {
          "default_scenes_per_run": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Scenes Per Run"
          },
          "default_prompt_art_style_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "random_mix",
                  "single_style"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Prompt Art Style Mode"
          },
          "default_prompt_art_style_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Prompt Art Style Text"
          },
          "social_posting_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Social Posting Enabled"
          }
        },
        "type": "object",
        "title": "AppSettingsUpdateRequest",
        "description": "Patch payload for updating global defaults."
      },
...
      "ArtStyleListsRead": {
        "properties": {
          "recommended_styles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Styles"
          },
          "other_styles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Other Styles"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "recommended_styles",
          "other_styles",
          "updated_at"
        ],
        "title": "ArtStyleListsRead",
        "description": "Line-oriented art-style settings payload split by style pool."
      },
      "ArtStyleListsUpdateRequest": {
        "properties": {
          "recommended_styles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Styles"
          },
          "other_styles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Other Styles"
          }
        },
        "type": "object",
        "required": [
          "recommended_styles",
          "other_styles"
        ],
        "title": "ArtStyleListsUpdateRequest",
        "description": "Full-replacement payload for editable style pools."
      },
      "ArtStyleRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_recommended": {
            "type": "boolean",
            "title": "Is Recommended"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "display_name",
          "description",
          "is_recommended",
          "is_active",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "title": "ArtStyleRead",
        "description": "Detailed representation of an art style entry."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}